Adding new files and modified vagrant file and bootstrap.sh

+++ centos7/bootstrap.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
-TELEGRAF_VERSION=telegraf-1.3.2-1.x86_64.rpm
-INFLUX_VERSION=influxdb-1.2.4.x86_64.rpm
-CHRONO_VERSION=chronograf-1.3.3.2.x86_64.rpm
+TELEGRAF_VERSION=telegraf-1.3.5-1.x86_64.rpm
+INFLUX_VERSION=influxdb-1.3.1.x86_64.rpm
+CHRONO_VERSION=chronograf-1.3.5.0.x86_64.rpm
 KAPACITOR_VERSION=kapacitor-1.3.1.x86_64.rpm
 
 cat <<EOT >> /etc/sysctl.conf
@@ -11,8 +11,10 @@ EOT
 sysctl -p
 
 systemctl restart network.service
+yum -y install epel-release
+yum repolist
 yum update -y
-yum install -y wget
+yum install -y wget nano
 
 # Install Influx DB
 wget -nv -O $INFLUX_VERSION https://dl.influxdata.com/influxdb/releases/$INFLUX_VERSION
@@ -22,7 +24,10 @@ systemctl start influxdb
 # Install Telegraf
 wget -nv -O $TELEGRAF_VERSION https://dl.influxdata.com/telegraf/releases/$TELEGRAF_VERSION
 yum localinstall -y $TELEGRAF_VERSION
-mv /vagrant/telegraf/telegraf.conf /etc/telegraf
+if [ ! -f /vagrant/telegraf/telegraf.conf ]; then
+    echo "Found telegraf.conf.  Installing."
+	mv /vagrant/telegraf/telegraf.conf /etc/telegraf
+fi
 systemctl start telegraf
 
 # Install Kapacitor
@@ -34,3 +39,7 @@ systemctl start kapacitor
 wget -nv -O $CHRONO_VERSION https://dl.influxdata.com/chronograf/releases/$CHRONO_VERSION
 yum localinstall -y $CHRONO_VERSION
 systemctl start chronograf
+
+# Install NodeJS
+curl --silent --location https://rpm.nodesource.com/setup_7.x | bash -
+yum -y install nodejs
